Board: w,dw,s are floats not int

+++ core/src/ch/asynk/rustanddust/engine/Board.java
@@ -38,9 +38,9 @@ public abstract class Board implements Disposable, Animation
         public int rows;
         public int x0;          // bottom left x offset
         public int y0;          // bottom left y offset
-        public int w;           // hex width
-        public int dw;          // half hex : w/2
-        public int s;           // hex side
+        public float w;         // hex width
+        public float dw;        // half hex : w/2
+        public float s;         // hex side
         public float dh;        // hex top : s/2
         public float h;         // square height : s + dh
         public float slope;     // north-west side slope : (dh / (float) dw)
